Add test, refactor some of the code
