+++ projects/fhi-angular-highcharts/src/lib/fhi-diagram-type-navs/fhi-diagram-type-nav.component.html
@@ -1,4 +1,4 @@
-<div class="fhi-diagram-type-nav mb-3" *ngIf="navDiagramTypeGroupsOk">
+<div class="fhi-diagram-type-nav mb-3" *ngIf="showNav">
   <div class="btn-group fhi-btn-group mb-1"
        role="group"
        aria-label="Velg tabell, kart eller graf">
