+++ projects/fhi-angular-highcharts/src/lib/fhi-diagram-type-navs/fhi-diagram-type-nav.models.ts
@@ -1,0 +1,9 @@
+import { FhiDiagramType } from '../fhi-diagram/fhi-diagram.models';
+
+export interface NavDiagramTypeGroup {
+  diagramType: FhiDiagramType;
+  icon: string;
+  id: string;
+  isDisabled: boolean;
+  name: string;
+}
