--- projects/fhi-angular-highcharts/src/lib/fhi-diagram-type-navs/fhi-diagram-type-navs.ts
@@ -1,3 +1,0 @@
-export class FhiDiagramTypeNavs {
-  static default = 'default';
-};
