+++ projects/fhi-angular-highcharts/src/public-api.ts
@@ -7,4 +7,4 @@ export * from './lib/fhi-angular-highcharts.module';
 
 export * from './lib/fhi-diagram/fhi-diagram.models';
 export * from './lib/fhi-diagram/fhi-diagram-types';
-export * from './lib/fhi-diagram-type-navs/fhi-diagram-type-navs';
+export * from './lib/fhi-diagram-type-navs/fhi-diagram-type-nav.constants';
